Implement HttpRequestHandler for handling HTTP requests and refactor TimetableServerCore to utilize it
